Refactor the ExpressionVisitor into its own file
