+++ UploadFIG/ObjectValidator.cs
@@ -9,16 +9,6 @@ using Hl7.FhirPath.Sprache;
 
 namespace UploadFIG
 {
-    public class FhirPathExpressionVisitor : BaseFhirPathExpressionVisitor
-    {
-        public FhirPathExpressionVisitor()
-            : base(ModelInspector.ForAssembly(typeof(Patient).Assembly),
-                  Hl7.Fhir.Model.ModelInfo.SupportedResources,
-                  Hl7.Fhir.Model.ModelInfo.OpenTypes)
-        {
-        }
-    }
-
     internal class ExpressionValidator
     {
         private readonly ModelInspector _mi = ModelInspector.ForAssembly(typeof(Patient).Assembly);
